pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = "Helen is a good english teacher"
authors = ["정승현 <tmdgusya@naver.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
langchain = "^0.0.129"
python-dotenv = "^1.0.0"
openai = "^0.27.2"
nltk = "^3.8.1"
pyaudio = "^0.2.13"
pydub = "^0.25.1"
pytest = "^7.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
